Update CAPS component and project-hub service to change references of CAPS to "Environmental Impact (CAPS)"
The diff includes changes to the CAPS component's HTML and TypeScript files to update the title of the CAPS page and the confirmation message when switching to the CAPS page from other pages. Additionally, the project-hub service's TypeScript file has been updated to change the title of the CAPS link in the side menu to "Environmental Impact (CAPS)".