feat: Add getDetailedScheduleData and putDetailedScheduleData methods to ProjectApiService
This commit adds two new methods, getDetailedScheduleData and putDetailedScheduleData, to the ProjectApiService in order to handle retrieving and updating detailed schedule data for a project. These methods make use of the Bryntum Gantt module and are necessary for enhancing the functionality of the detailed schedule component in the project hub.